refactor vo directory
